Use something else then port 8080 and 8081
The operator cannot be run locally when something else runs on 8080
(a very common port). In particular, on a system with the assisted
installer, there is a crucial service running on 8080. The proposal
is to change it to anything else than 8080. It's still possible to
override the port using the commandline, but this change reduces the
odds o having a conflict by default